reworked composites and some other tweaking
