--- DearPyGui/src/core/AppItems/composite/mvSourceWindow.h
@@ -1,31 +1,0 @@
-#pragma once
-#include "mvTextEditor.h"
-#include "mvTypeBases.h"
-
-namespace Marvel {
-
-	class mvSourceWindow : public mvBaseWindowAppitem
-	{
-
-		MV_APPLY_WIDGET_REGISTRATION_OLD_SYSTEM(mvAppItemType::SourceWindow, "add_source_window")
-
-	public:
-
-		mvSourceWindow(const std::string& name) : 
-			mvBaseWindowAppitem(name)
-		{
-			m_windowflags = ImGuiWindowFlags_HorizontalScrollbar | ImGuiWindowFlags_MenuBar;
-		}
-
-		void setFile(const std::string& file);
-
-		void draw(ImDrawList* drawlist, float x, float y) override;
-
-	private:
-
-		mvTextEditor  m_editor;
-		std::string  m_file;
-
-	};
-
-}
