+++ DearPyGui/src/core/AppItems/mvTypeBases.cpp
@@ -500,118 +500,4 @@ namespace Marvel {
 		return false;
 	}
 
-	mvBaseWindowAppitem::mvBaseWindowAppitem(const std::string& name)
-		: 
-		mvAppItem(name)
-	{
-		m_width = 500;
-		m_height = 500;
-	}
-
-	void mvBaseWindowAppitem::addFlag(ImGuiWindowFlags flag) 
-	{ 
-		m_windowflags |= flag; 
-	}
-
-	void mvBaseWindowAppitem::removeFlag(ImGuiWindowFlags flag) 
-	{ 
-		m_windowflags &= ~flag; 
-	}
-
-	void mvBaseWindowAppitem::setWidth(int width)
-	{ 
-		m_width = width;
-		m_dirty_size = true; 
-	}
-
-	void mvBaseWindowAppitem::setHeight(int height)
-	{
-		m_height = height;
-		m_dirty_size = true; 
-	}
-
-	bool mvBaseWindowAppitem::prerender()
-	{
-
-		if (!m_show)
-			return false;
-
-		if (m_dirty_size)
-		{
-			ImGui::SetNextWindowSize(ImVec2((float)m_width, (float)m_height));
-			m_dirty_size = false;
-		}
-
-		if (m_dirtyPos)
-		{
-			ImGui::SetNextWindowPos(m_state.getItemPos());
-			m_dirtyPos = false;
-		}
-
-		if (m_focusNextFrame)
-		{
-			ImGui::SetNextWindowFocus();
-			m_focusNextFrame = false;
-		}
-
-		if (!ImGui::Begin(m_label.c_str(), &m_show, m_windowflags))
-		{
-			ImGui::End();
-			return false;
-		}
-
-		return true;
-	}
-
-	void mvBaseWindowAppitem::handleSpecificKeywordArgs(PyObject* dict)
-	{
-		if (dict == nullptr)
-			return;
-
-		// helper for bit flipping
-		auto flagop = [dict](const char* keyword, int flag, int& flags)
-		{
-			if (PyObject* item = PyDict_GetItemString(dict, keyword)) ToBool(item) ? flags |= flag : flags &= ~flag;
-		};
-
-		// window flags
-		flagop("autosize", ImGuiWindowFlags_AlwaysAutoResize, m_windowflags);
-		flagop("no_move", ImGuiWindowFlags_NoMove, m_windowflags);
-		flagop("no_resize", ImGuiWindowFlags_NoResize, m_windowflags);
-		flagop("no_title_bar", ImGuiWindowFlags_NoTitleBar, m_windowflags);
-		flagop("no_scrollbar", ImGuiWindowFlags_NoScrollbar, m_windowflags);
-		flagop("no_collapse", ImGuiWindowFlags_NoCollapse, m_windowflags);
-		flagop("horizontal_scrollbar", ImGuiWindowFlags_HorizontalScrollbar, m_windowflags);
-		flagop("no_focus_on_appearing", ImGuiWindowFlags_NoFocusOnAppearing, m_windowflags);
-		flagop("no_bring_to_front_on_focus", ImGuiWindowFlags_NoBringToFrontOnFocus, m_windowflags);
-		flagop("menubar", ImGuiWindowFlags_MenuBar, m_windowflags);
-		flagop("no_background", ImGuiWindowFlags_NoBackground, m_windowflags);
-
-	}
-
-	void mvBaseWindowAppitem::getSpecificConfiguration(PyObject* dict)
-	{
-		if (dict == nullptr)
-			return;
-
-		// helper to check and set bit
-		auto checkbitset = [dict](const char* keyword, int flag, const int& flags)
-		{
-			PyDict_SetItemString(dict, keyword, ToPyBool(flags & flag));
-		};
-
-		// window flags
-		checkbitset("autosize", ImGuiWindowFlags_AlwaysAutoResize, m_windowflags);
-		checkbitset("no_resize", ImGuiWindowFlags_NoResize, m_windowflags);
-		checkbitset("no_title_bar", ImGuiWindowFlags_NoTitleBar, m_windowflags);
-		checkbitset("no_move", ImGuiWindowFlags_NoMove, m_windowflags);
-		checkbitset("no_scrollbar", ImGuiWindowFlags_NoScrollbar, m_windowflags);
-		checkbitset("no_collapse", ImGuiWindowFlags_NoCollapse, m_windowflags);
-		checkbitset("horizontal_scrollbar", ImGuiWindowFlags_HorizontalScrollbar, m_windowflags);
-		checkbitset("no_focus_on_appearing", ImGuiWindowFlags_NoFocusOnAppearing, m_windowflags);
-		checkbitset("no_bring_to_front_on_focus", ImGuiWindowFlags_NoBringToFrontOnFocus, m_windowflags);
-		checkbitset("menubar", ImGuiWindowFlags_MenuBar, m_windowflags);
-		checkbitset("no_background", ImGuiWindowFlags_NoBackground, m_windowflags);
-	}
-
 }
