+++ DearPyGui/src/core/AppItems/textures/mvTextureContainer.h
@@ -6,7 +6,7 @@
 
 namespace Marvel {
 
-    MV_REGISTER_WIDGET(mvTextureContainer, MV_ITEM_DESC_ROOT, StorageValueTypes::None, 1);
+    MV_REGISTER_WIDGET(mvTextureContainer, MV_ITEM_DESC_ROOT | MV_ITEM_DESC_CONTAINER, StorageValueTypes::None, 1);
     class mvTextureContainer : public mvAppItem
     {
