+++ DearPyGui/src/core/Theming/mvImGuiThemeScope.h
@@ -3,7 +3,7 @@
 #include <unordered_map>
 #include "mvEvents.h"
 #include "mvAppItem.h"
-#include "composite/mvStyleWindow.h"
+#include "mvStyleWindow.h"
 #include "mvThemeManager.h"
 
 namespace Marvel {
