+++ DearPyGui/src/core/Theming/mvImNodesThemeScope.h
@@ -3,7 +3,7 @@
 #include <unordered_map>
 #include "mvEvents.h"
 #include "mvAppItem.h"
-#include "composite/mvStyleWindow.h"
+#include "mvStyleWindow.h"
 #include "mvThemeManager.h"
 #include <imnodes.h>
