+++ DearPyGui/src/core/Tooling/mvDocWindow.cpp
@@ -1,43 +1,9 @@
+#include <map>
+#include "mvPythonParser.h"
 #include "mvDocWindow.h"
-#include <imgui.h>
-#include "mvApp.h"
 #include "mvModule_Core.h"
-#include "mvInput.h"
-#include "mvItemRegistry.h"
-#include "mvFontScope.h"
 
 namespace Marvel {
-
-	void mvDocWindow::InsertParser(std::map<std::string, mvPythonParser>* parsers)
-	{
-		mvPythonParser parser(mvPyDataType::String);
-		mvAppItem::AddCommonArgs(parser);
-		parser.removeArg("parent");
-		parser.removeArg("before");
-		parser.removeArg("source");
-		parser.removeArg("callback");
-		parser.removeArg("callback_data");
-		parser.removeArg("enabled");
-
-		parser.addArg<mvPyDataType::Integer>("x_pos", mvArgType::KEYWORD_ARG, "200");
-		parser.addArg<mvPyDataType::Integer>("y_pos", mvArgType::KEYWORD_ARG, "200");
-
-		parser.addArg<mvPyDataType::Bool>("autosize", mvArgType::KEYWORD_ARG, "False", "Autosized the window to fit it's items.");
-		parser.addArg<mvPyDataType::Bool>("no_resize", mvArgType::KEYWORD_ARG, "False", "Allows for the window size to be changed or fixed");
-		parser.addArg<mvPyDataType::Bool>("no_title_bar", mvArgType::KEYWORD_ARG, "False", "Title name for the title bar of the window");
-		parser.addArg<mvPyDataType::Bool>("no_move", mvArgType::KEYWORD_ARG, "False", "Allows for the window's position to be changed or fixed");
-		parser.addArg<mvPyDataType::Bool>("no_scrollbar", mvArgType::KEYWORD_ARG, "False", " Disable scrollbars (window can still scroll with mouse or programmatically)");
-		parser.addArg<mvPyDataType::Bool>("no_collapse", mvArgType::KEYWORD_ARG, "False", "Disable user collapsing window by double-clicking on it");
-		parser.addArg<mvPyDataType::Bool>("horizontal_scrollbar", mvArgType::KEYWORD_ARG, "False", "Allow horizontal scrollbar to appear (off by default).");
-		parser.addArg<mvPyDataType::Bool>("no_focus_on_appearing", mvArgType::KEYWORD_ARG, "False", "Disable taking focus when transitioning from hidden to visible state");
-		parser.addArg<mvPyDataType::Bool>("no_bring_to_front_on_focus", mvArgType::KEYWORD_ARG, "False", "Disable bringing window to front when taking focus (e.g. clicking on it or programmatically giving it focus)");
-		parser.addArg<mvPyDataType::Bool>("no_close", mvArgType::KEYWORD_ARG, "False");
-		parser.addArg<mvPyDataType::Bool>("no_background", mvArgType::KEYWORD_ARG, "False");
-
-		parser.finalize();
-
-		parsers->insert({ s_command, parser });
-	}
 
 	static void ColorText(const char* item)
 	{
@@ -59,13 +25,13 @@ namespace Marvel {
 		ImGui::TextColored(ImGui::GetStyleColorVec4(ImGuiCol_Text), desc); ImGui::NextColumn();
 	}
 
-	mvDocWindow::mvDocWindow(const std::string& name)
-		: mvBaseWindowAppitem(name)
-	{
+    mvDocWindow::mvDocWindow()
+    {
+        m_windowflags = ImGuiWindowFlags_NoSavedSettings;
 		m_width = 700;
 		m_height = 500;
 		setup();
-	}
+    }
 
 	void mvDocWindow::setup()
 	{
@@ -77,7 +43,7 @@ namespace Marvel {
 		{
 			m_constantsValues.emplace_back("Constant with a value of " + std::to_string(item.second));
 			m_docCategories["Constants"].push_back(m_constantsValues.back().c_str());
-			m_commandCategories["Constants"].push_back(item.first.c_str());		
+			m_commandCategories["Constants"].push_back(item.first.c_str());
 		}
 
 		for (const auto& item : docmap)
@@ -111,19 +77,14 @@ namespace Marvel {
 		}
 
 		m_categories.push_back("Constants");
-		
+
 	}
 
-	void mvDocWindow::draw(ImDrawList* drawlist, float x, float y)
-	{
-		mvFontScope fscope(this);
-
-		if (!prerender())
-			return;
+    void mvDocWindow::drawWidgets()
+    {
 
 		if (ImGui::BeginTabBar("Main Tabbar##doc"))
 		{
-
 			if (ImGui::BeginTabItem("Commands##doc"))
 			{
 				ImGui::BeginGroup();
@@ -199,24 +160,6 @@ namespace Marvel {
 
 			ImGui::EndTabBar();
 		}
-
-		if (ImGui::IsWindowFocused())
-		{
-
-			float titleBarHeight = ImGui::GetStyle().FramePadding.y * 2 + ImGui::GetFontSize();
-
-			// update mouse
-			ImVec2 mousePos = ImGui::GetMousePos();
-			float x = mousePos.x - ImGui::GetWindowPos().x;
-			float y = mousePos.y - ImGui::GetWindowPos().y - titleBarHeight;
-			mvInput::setMousePosition(x, y);
-
-			if (mvApp::GetApp()->getItemRegistry().getActiveWindow() != "documentation##standard")
-				mvEventBus::Publish(mvEVT_CATEGORY_ITEM, mvEVT_ACTIVE_WINDOW, { CreateEventArgument("WINDOW", std::string("documentation##standard")) });
-
-		}
-
-		ImGui::End();
-	}
+    }
 
 }
