+++ DearPyGui/src/core/Tooling/mvStyleWindow.cpp
@@ -1,11 +1,6 @@
-#include <imgui.h>
-#include "mvApp.h"
-#include "mvAppItemCommons.h"
 #include "mvStyleWindow.h"
-#include "mvInput.h"
-#include "mvItemRegistry.h"
-#include "mvImGuiThemeScope.h"
-#include "mvFontScope.h"
+#include "mvApp.h"
+#include "mvThemeManager.h"
 
 // Helper to display a little (?) mark which shows a tooltip when hovered.
 // In your own code you may want to display an actual icon if you are using a merged icon fonts (see docs/FONTS.md)
@@ -24,49 +19,19 @@ static void HelpMarker(const char* desc)
 
 namespace Marvel {
 
-    void mvStyleWindow::InsertParser(std::map<std::string, mvPythonParser>* parsers)
+
+    mvStyleWindow::mvStyleWindow()
     {
-        mvPythonParser parser(mvPyDataType::String);
-        mvAppItem::AddCommonArgs(parser);
-        parser.removeArg("parent");
-        parser.removeArg("before");
-        parser.removeArg("source");
-        parser.removeArg("callback");
-        parser.removeArg("callback_data");
-        parser.removeArg("enabled");
-
-        parser.addArg<mvPyDataType::Integer>("x_pos", mvArgType::KEYWORD_ARG, "200");
-        parser.addArg<mvPyDataType::Integer>("y_pos", mvArgType::KEYWORD_ARG, "200");
-
-        parser.addArg<mvPyDataType::Bool>("autosize", mvArgType::KEYWORD_ARG, "False", "Autosized the window to fit it's items.");
-        parser.addArg<mvPyDataType::Bool>("no_resize", mvArgType::KEYWORD_ARG, "False", "Allows for the window size to be changed or fixed");
-        parser.addArg<mvPyDataType::Bool>("no_title_bar", mvArgType::KEYWORD_ARG, "False", "Title name for the title bar of the window");
-        parser.addArg<mvPyDataType::Bool>("no_move", mvArgType::KEYWORD_ARG, "False", "Allows for the window's position to be changed or fixed");
-        parser.addArg<mvPyDataType::Bool>("no_scrollbar", mvArgType::KEYWORD_ARG, "False", " Disable scrollbars (window can still scroll with mouse or programmatically)");
-        parser.addArg<mvPyDataType::Bool>("no_collapse", mvArgType::KEYWORD_ARG, "False", "Disable user collapsing window by double-clicking on it");
-        parser.addArg<mvPyDataType::Bool>("horizontal_scrollbar", mvArgType::KEYWORD_ARG, "False", "Allow horizontal scrollbar to appear (off by default).");
-        parser.addArg<mvPyDataType::Bool>("no_focus_on_appearing", mvArgType::KEYWORD_ARG, "False", "Disable taking focus when transitioning from hidden to visible state");
-        parser.addArg<mvPyDataType::Bool>("no_bring_to_front_on_focus", mvArgType::KEYWORD_ARG, "False", "Disable bringing window to front when taking focus (e.g. clicking on it or programmatically giving it focus)");
-        parser.addArg<mvPyDataType::Bool>("no_close", mvArgType::KEYWORD_ARG, "False");
-        parser.addArg<mvPyDataType::Bool>("no_background", mvArgType::KEYWORD_ARG, "False");
-
-        parser.finalize();
-
-        parsers->insert({ s_command, parser });
+        m_windowflags = ImGuiWindowFlags_NoSavedSettings;
     }
 
-    void mvStyleWindow::draw(ImDrawList* drawlist, float x, float y)
+    void mvStyleWindow::drawWidgets()
     {
-
-        mvFontScope fscope(this);
-
-        if (!prerender())
-            return;
 
         ImGui::BeginChild("##colors", ImVec2(0, 0), true, ImGuiWindowFlags_AlwaysVerticalScrollbar | ImGuiWindowFlags_AlwaysHorizontalScrollbar | ImGuiWindowFlags_NavFlattened);
         ImGui::PushItemWidth(-350);
 
-        if(ImGui::BeginTabBar("##tabs", ImGuiTabBarFlags_None))
+        if (ImGui::BeginTabBar("##tabs", ImGuiTabBarFlags_None))
         {
 
             if (ImGui::BeginTabItem("Styles"))
@@ -103,7 +68,7 @@ namespace Marvel {
                         continue;
 
                     ImGui::PushID(&item);
-                    if(ImGui::SliderFloat("##style", std::get<2>(item), 0.0f, std::get<3>(item)))
+                    if (ImGui::SliderFloat("##style", std::get<2>(item), 0.0f, std::get<3>(item)))
                         mvThemeManager::InValidateStyleTheme();
                     ImGui::SameLine();
                     ImGui::TextUnformatted(std::get<0>(item).c_str());
@@ -188,7 +153,7 @@ namespace Marvel {
                         continue;
 
                     ImGui::PushID(&item);
-                    if(ImGui::ColorEdit4("##color(disable)", *std::get<2>(item), ImGuiColorEditFlags_AlphaBar))
+                    if (ImGui::ColorEdit4("##color(disable)", *std::get<2>(item), ImGuiColorEditFlags_AlphaBar))
                         mvThemeManager::InValidateDisabledColorTheme();
                     ImGui::SameLine();
                     ImGui::TextUnformatted(std::get<0>(item).c_str());
@@ -205,23 +170,6 @@ namespace Marvel {
         ImGui::PopItemWidth();
         ImGui::EndChild();
 
-        if (ImGui::IsWindowFocused())
-        {
-
-            float titleBarHeight = ImGui::GetStyle().FramePadding.y * 2 + ImGui::GetFontSize();
-
-            // update mouse
-            ImVec2 mousePos = ImGui::GetMousePos();
-            float x = mousePos.x - ImGui::GetWindowPos().x;
-            float y = mousePos.y - ImGui::GetWindowPos().y - titleBarHeight;
-            mvInput::setMousePosition(x, y);
-
-            if (mvApp::GetApp()->getItemRegistry().getActiveWindow() != "style##standard")
-                mvEventBus::Publish(mvEVT_CATEGORY_ITEM, mvEVT_ACTIVE_WINDOW, { CreateEventArgument("WINDOW", std::string("style##standard")) });
-
-        }
-
-        ImGui::End();
     }
 
 }
